requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pymongo[srv]==3.12
pydantic==2.11.3
python-dotenv==1.1.0
